refactor(commands): extract config tests to sibling file
